add sharing tests
