+++ app/modules/sharing/_styles.scss
@@ -3,6 +3,6 @@
   border: 0;
 
   & *:hover {
-    color: $mt_color-neutral-l;
+    color: $mt2_color-neutral-l;
   }
 }
