+++ app/modules/sharing/scripts/MTSharing.jsx
@@ -3,84 +3,12 @@
 import React, { Component, PropTypes }  from 'react';
 import {Pestle} from '@natgeo/mortar-pestle';
 import urlEncode from 'urlencode';
-
-class MTSocialButton extends Component {
-  constructor(props) {
-    super(props);
-    this.handleClick = this.handleClick.bind(this);
-  }
-
-  openShareWindow(site, url, title) {
-    const encodedUrl = urlEncode(url);
-    const siteAttr = {
-      facebook: {
-        h: 279,
-        url: "http://www.facebook.com/share.php?u=" + encodedUrl,
-        w: 575,
-        windowTitle: 'Facebook'
-      },
-      'google-plus': {
-        h: 647,
-        url: "https://plus.google.com/share?url=" + encodedUrl,
-        w: 500,
-        windowTitle: 'Google Plus'
-      },
-      linkedin: {
-        h: 425,
-        url: "https://www.linkedin.com/shareArticle?mini=true&url=nationalgeographic.com" + encodedUrl,
-        w: 450,
-        windowTitle: 'LinkedIn'
-      },
-      pinterest: {
-        h: 575,
-        url: "http://pinterest.com/pin/create/button/?" + encodedUrl,
-        w: 750,
-        windowTitle: 'Pinterest'
-      },
-      twitter: {
-        h: 420,
-        url: "https://twitter.com/intent/tweet?url="
-              + encodedUrl
-              + "&text="
-              + urlEncode(title)
-              + "&via=NatGeo",
-        w: 550,
-        windowTitle: 'Twitter'
-      }
-    }
-    const viewportWidth = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-    const viewportHeight = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-    const windowLeft = (viewportWidth - siteAttr[site].w) / 2;
-    const windowTop = (viewportHeight - siteAttr[site].h) / 2;
-
-    window.open(
-      siteAttr[site].url,
-      siteAttr[site].windowTitle,
-      "status = 1, width = " + siteAttr[site].w
-        + ", height = " + siteAttr[site].h
-        + ", left = " + windowLeft
-        + ", top = " + windowTop
-        + ", resizable = 0"
-      );
-  }
-
-  handleClick() {
-    this.openShareWindow(this.props.site, this.props.url, this.props.title);
-  }
-
-  render() {
-    return (
-      <button onClick={this.handleClick} className="mt2_sharing-btn">
-        <svg className={"mt2_icon mt2_color--social--" + this.props.site}>
-          <use xmlnsXlink="http://www.w3.org/1999/xlink" xlinkHref={"#social-" + this.props.site}></use>
-        </svg>
-      </button>
-    )
-  }
-}
+import MTSocialButton from './MTSocialButton.jsx'
 
 class MTSharing extends Component {
   render() {
+    // this looks wacky, but the strings need to be URL encoded
+    // while the API remains unencoded
     const emailFormat = 'mailto:?subject='
       + urlEncode(this.props.title)
       + '&body='
@@ -94,14 +22,14 @@ class MTSharing extends Component {
       )
     ;
 
-    let socialButtons = this.props.sharingOptions.map(option => {
+    const socialButtons = this.props.buttons.map(option => {
       if(option !== 'email') {
         return (
           <MTSocialButton
-          key={option}
-          url={this.props.url}
-          title={this.props.title}
-          site={option}
+            key={option}
+            url={this.props.url}
+            title={this.props.title}
+            site={option}
           />
         )
       }
@@ -123,7 +51,7 @@ class MTSharing extends Component {
   }
 }
 
-const sharingOptions = [
+const buttonDefaults = [
   'facebook',
   'twitter',
   'google-plus',
@@ -133,12 +61,12 @@ const sharingOptions = [
 ]
 
 MTSharing.defaultProps = {
-  sharingOptions: sharingOptions
+  buttons: buttonDefaults
 }
 
 MTSharing.propTypes = {
   url: PropTypes.string.isRequired,
-  sharingOptions: PropTypes.arrayOf(React.PropTypes.string),
+  buttons: PropTypes.arrayOf(React.PropTypes.string),
   title: PropTypes.string.isRequired
 }
